Console app (#2)
* Refactor cost computation logic and enhance console app

Removed ComputeCost method from CheckoutItem and SalesItem, and added it to PricingCatalogue to centralize cost computation. Enhanced the console app with a full user interaction loop for item purchasing and checkout, and added detailed purchase breakdown. Updated PricingRule and related entities to maintain consistency.

* Add PricingUnit to PricingInfo and refactor related tests.

Added PricingUnit to the PricingInfo constructor and adjusted existing methods and tests to accommodate this change. This includes updating the logic in ComputeCost methods and modifying how pricing rules are instantiated in unit tests.

---------

Co-authored-by: Rodger Gu <rodger.gu@celohealth.com>